scanned item pops up. need to work on changing quantity and retaining it on switch to item list

+++ display/src/elements/number_setting.cpp
@@ -17,9 +17,9 @@
  */
 NumberSetting::NumberSetting(struct DisplayGlobal displayGlobal,
                              const SDL_Rect& boundaryRectangle,
-                             int settingId,
-                             const std::string& logFile)
-    : settingId(settingId) {
+                             const std::string& logFile,
+                             const int settingId = -1)
+    : settingId(settingId), {
   setupPosition(boundaryRectangle);
   this->logger  = std::make_unique<Logger>(logFile);
   this->logFile = logFile;
@@ -42,6 +42,37 @@ NumberSetting::NumberSetting(struct DisplayGlobal displayGlobal,
   FoodItem foodItem  = readFoodItemById(this->settingId);
   this->settingValue = foodItem.getQuantity();
   this->children[1]->setContent(std::to_string(this->settingValue));
+}
+
+/**
+ * Get the current value of the setting from the database and set the content to it.
+ *
+ * @param displayGlobal
+ * @param boundaryRectangle Rectangle defining offset within parent (if any) and width +
+ * height
+ * @param settingId The primary key of the food item corresponding to this object
+ */
+NumberSetting::NumberSetting(struct DisplayGlobal displayGlobal,
+                             const SDL_Rect& boundaryRectangle,
+                             const std::string& logFile) {
+  setupPosition(boundaryRectangle);
+  this->logger  = std::make_unique<Logger>(logFile);
+  this->logFile = logFile;
+
+  std::shared_ptr<Button> decreaseButton = std::make_shared<Button>(
+      displayGlobal, SDL_Rect{0, 0, 0, 0}, "-", SDL_Point{0, 0},
+      [this]() { this->settingValue--; }, this->logFile);
+  addElement(std::move(decreaseButton));
+
+  std::shared_ptr<Text> text = std::make_shared<Text>(displayGlobal, SDL_Rect{0, 0, 0, 0},
+                                                      DisplayGlobal::futuramFontPath, "0",
+                                                      24, SDL_Color{0, 255, 0, 255});
+  addElement(std::move(text));
+
+  std::shared_ptr<Button> increaseButton = std::make_shared<Button>(
+      displayGlobal, SDL_Rect{0, 0, 0, 0}, "+", SDL_Point{0, 0},
+      [this]() { this->settingValue++; }, this->logFile);
+  addElement(std::move(increaseButton));
 }
 
 /**
@@ -75,16 +106,18 @@ void NumberSetting::updateSelf() {
     this->children[i]->setPositionRelativeToParent(childRelativePosition);
   }
 
-  FoodItem foodItem    = readFoodItemById(this->settingId);
-  int foodItemQuantity = foodItem.getQuantity();
-  if (foodItemQuantity == 0) {
-    deleteById(this->settingId);
-  }
-  else if (foodItemQuantity != this->settingValue) {
-    updateFoodItemQuantity(this->settingId, this->settingValue);
-  }
-  else {
-    this->children[1]->setContent(std::to_string(this->settingValue));
+  if (this->settingId != -1) {
+    FoodItem foodItem    = readFoodItemById(this->settingId);
+    int foodItemQuantity = foodItem.getQuantity();
+    if (foodItemQuantity == 0) {
+      deleteById(this->settingId);
+    }
+    else if (foodItemQuantity != this->settingValue) {
+      updateFoodItemQuantity(this->settingId, this->settingValue);
+    }
+    else {
+      this->children[1]->setContent(std::to_string(this->settingValue));
+    }
   }
 }
